+++ web/src/modules/Index/components/Shortcuts/index.tsx
@@ -2,10 +2,10 @@ import { Flex } from "@mantine/core";
 
 import Shortcut from "./components/Shortcut";
 import classes from "./styles.module.scss";
-import { useShortcutsStore } from "@store/shortcuts";
+import { useStatrpageStore } from "@store/startpage";
 
 export const Shortcuts = () => {
-  const shortcuts = useShortcutsStore((state) => state.shortcuts);
+  const shortcuts = useStatrpageStore((state) => state.shortcuts);
 
   const items = shortcuts.map((item, i) => {
     return <Shortcut key={i} idx={i} {...item} />;