+++ web/src/modules/Settings/components/_startpage/ToDo/components/ToDoSwitch.tsx
@@ -1,0 +1,13 @@
+import { Switch } from "@mantine/core";
+import { useStatrpageStore } from "@store/startpage";
+
+const ToDoSwitch = () => {
+  const displayTodos = useStatrpageStore((state) => state.displayTodos);
+  const setDisplayTodos = useStatrpageStore((state) => state.setDisplayTodos);
+
+  return (
+    <Switch checked={displayTodos} onChange={(e) => setDisplayTodos(e.currentTarget.checked)} />
+  );
+};
+
+export default ToDoSwitch;