+++ web/src/store/startpage/index.tsx
@@ -1,20 +1,5 @@
-import { IFaviconAPI } from "@store/settings";
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
-
-export type IAutocompleteEngines = "google" | "duckduckgo" | "brave" | "qwant";
-
-export type ICategories =
-  | "general"
-  | "images"
-  | "videos"
-  | "news"
-  | "maps"
-  | "music"
-  | "it"
-  | "science"
-  | "files"
-  | "social_media";
 
 export interface IShortcut {
   title: string;
@@ -22,32 +7,49 @@ export interface IShortcut {
   imgUrl?: string;
 }
 
-interface ShortcutsState {
+export interface IToDo {
+  label: string;
+  checked: string;
+}
+
+interface StartpageState {
   hydrated: boolean;
 
   shortcuts: IShortcut[];
   setShortcuts: (next: IShortcut[]) => void;
-
   displayShortcuts: boolean;
   setDisplayShortcuts: (next: boolean) => void;
-
   openInNewTab: boolean;
   setOpenInNewTab: (next: boolean) => void;
+
+  todos: IToDo[];
+  setTodos: (next: IToDo[]) => void;
+  displayTodos: boolean;
+  setDisplayTodos: (next: boolean) => void;
+
+  displayWeather: boolean;
+  setDislpayWeather: (next: boolean) => void;
 }
 
-export const useShortcutsStore = create<ShortcutsState>()(
+export const useStatrpageStore = create<StartpageState>()(
   persist(
     (set) => ({
       hydrated: false,
 
       shortcuts: [],
       setShortcuts: (next) => set({ shortcuts: next }),
-
       displayShortcuts: false,
       setDisplayShortcuts: (next) => set({ displayShortcuts: next }),
-
       openInNewTab: false,
       setOpenInNewTab: (next) => set({ openInNewTab: next }),
+
+      todos: [],
+      setTodos: (next) => set({ todos: next }),
+      displayTodos: false,
+      setDisplayTodos: (next) => set({ displayTodos: next }),
+
+      displayWeather: false,
+      setDislpayWeather: (next) => set({ displayWeather: next }),
     }),
     {
       onRehydrateStorage: () => (state) => {
@@ -61,6 +63,9 @@ export const useShortcutsStore = create<ShortcutsState>()(
         displayShortcuts: state.displayShortcuts,
         shortcuts: state.shortcuts,
         openInNewTab: state.openInNewTab,
+        todos: state.todos,
+        displayTodos: state.displayTodos,
+        displayWeather: state.displayWeather,
       }),
     }
   )