+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Instant answers - TicTacToe, Dice Roll
 - Open Meteo as weather data source
 - Prioritize or blacklist certain websites
+- More start page customization
 
 ### Fixed